LayoutTests/editing/secure-input/password-input-focusing-expected.txt

Verify that basic focusing/unfocusing updates secure input state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial state, no selection:
PASS testRunner.secureEventInputIsEnabled is false

A password input is focused:
PASS testRunner.secureEventInputIsEnabled is true

A regular text input is focused:
PASS testRunner.secureEventInputIsEnabled is false

A password input is focused again:
PASS testRunner.secureEventInputIsEnabled is true
PASS successfullyParsed is true

TEST COMPLETE
